Moved get pipeline instance api to pipeline instance controller
