+++ api/api-pipeline-instance-v1/src/main/java/com/thoughtworks/go/apiv1/pipelineinstance/representers/MaterialRepresenter.java
@@ -14,18 +14,17 @@
  * limitations under the License.
  */
 
-package com.thoughtworks.go.apiv1.pipelineoperations.representers;
+package com.thoughtworks.go.apiv1.pipelineinstance.representers;
 
 import com.thoughtworks.go.api.base.OutputWriter;
-import com.thoughtworks.go.domain.materials.Modification;
+import com.thoughtworks.go.domain.materials.Material;
 
-public class ModificationRepresenter {
-    public static void toJSON(OutputWriter outputWriter, Modification modification) {
-        outputWriter.add("id", modification.getId())
-                .add("revision", modification.getRevision())
-                .addIfNotNull("modified_time", modification.getModifiedTime())
-                .add("user_name", modification.getUserName())
-                .add("comment", modification.getComment())
-                .add("email_address", modification.getEmailAddress());
+public class MaterialRepresenter {
+    public static void toJSON(OutputWriter outputWriter, Material material) {
+        outputWriter.add("id", material.getId())
+                .add("name", material.getDisplayName())
+                .add("fingerprint", material.getFingerprint())
+                .add("type", material.getTypeForDisplay())
+                .add("description", material.getLongDescription());
     }
 }
