+++ api/api-pipeline-instance-v1/src/main/java/com/thoughtworks/go/apiv1/pipelineinstance/representers/MaterialRevisionRepresenter.java
@@ -13,17 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.thoughtworks.go.apiv1.pipelineoperations.representers;
+package com.thoughtworks.go.apiv1.pipelineinstance.representers;
 
 import com.thoughtworks.go.api.base.OutputWriter;
-import com.thoughtworks.go.api.representers.JsonReader;
 import com.thoughtworks.go.domain.MaterialRevision;
-import com.thoughtworks.go.server.domain.MaterialForScheduling;
 
 public class MaterialRevisionRepresenter {
-    public static MaterialForScheduling fromJSON(JsonReader jsonReader) {
-        String name = jsonReader.getString("fingerprint");
-        String value = jsonReader.getString("revision");
-        return new MaterialForScheduling(name, value);
+    public static void toJSON(OutputWriter outputWriter, MaterialRevision revision) {
+        outputWriter
+                .add("changed", revision.isChanged())
+                .addChild("material", materialWriter -> MaterialRepresenter.toJSON(outputWriter, revision.getMaterial()))
+                .addChildList("modifications", modificationWriter -> revision.getModifications()
+                        .forEach(modification -> modificationWriter.addChild(modWriter -> ModificationRepresenter.toJSON(modWriter, modification))));
     }
 }
