+++ api/api-pipeline-instance-v1/src/test/groovy/com/thoughtworks/go/apiv1/pipelineinstance/representers/MaterialRepresenterTest.groovy
@@ -14,34 +14,28 @@
  * limitations under the License.
  */
 
-package com.thoughtworks.go.apiv1.pipelineoperations.representers
+package com.thoughtworks.go.apiv1.pipelineinstance.representers
 
-import com.thoughtworks.go.domain.buildcause.BuildCause
-import com.thoughtworks.go.helper.ModificationsMother
+import com.thoughtworks.go.helper.MaterialsMother
 import org.junit.jupiter.api.Test
 
-import static com.thoughtworks.go.api.base.JsonUtils.toObject
 import static com.thoughtworks.go.api.base.JsonUtils.toObjectString
 import static net.javacrumbs.jsonunit.fluent.JsonFluentAssert.assertThatJson
 
-class BuildCauseRepresenterTest {
+class MaterialRepresenterTest {
   @Test
-  void 'should serialize build cause to json'() {
-    def materialRevisions = ModificationsMother.multipleModifications()
-    def buildCause = BuildCause.createWithModifications(materialRevisions, "approver")
-
-    def actualJson = toObjectString({ BuildCauseRepresenter.toJSON(it, buildCause) })
+  void 'should render material config properties'() {
+    def material = MaterialsMother.gitMaterial("git1")
 
     def expectedJSON = [
-      "trigger_message"   : "modified by committer <html />",
-      "trigger_forced"    : false,
-      "approver"          : "approver",
-      "material_revisions": buildCause.getMaterialRevisions().collect { eachItem ->
-        toObject({
-          MaterialRevisionRepresenter.toJSON(it, eachItem)
-        })
-      }
+      id         : -1,
+      name       : "git1",
+      fingerprint: "2fde537a026695884e2ee13e8f9730eca0610a3e407dbcc6bbce974f595c2f7c",
+      type       : "Git",
+      description: "URL: git1, Branch: master"
     ]
+
+    def actualJson = toObjectString({ MaterialRepresenter.toJSON(it, material) })
 
     assertThatJson(actualJson).isEqualTo(expectedJSON)
   }
