+++ api/api-pipeline-operations-v1/src/main/java/com/thoughtworks/go/apiv1/pipelineoperations/PipelineOperationsControllerV1.java
@@ -21,7 +21,6 @@ import com.thoughtworks.go.api.ApiVersion;
 import com.thoughtworks.go.api.representers.JsonReader;
 import com.thoughtworks.go.api.spring.ApiAuthenticationHelper;
 import com.thoughtworks.go.api.util.GsonTransformer;
-import com.thoughtworks.go.apiv1.pipelineoperations.representers.PipelineInstanceModelRepresenter;
 import com.thoughtworks.go.apiv1.pipelineoperations.representers.PipelineScheduleOptionsRepresenter;
 import com.thoughtworks.go.apiv1.pipelineoperations.representers.TriggerOptions;
 import com.thoughtworks.go.apiv1.pipelineoperations.representers.TriggerWithOptionsViewRepresenter;
@@ -82,14 +81,12 @@ public class PipelineOperationsControllerV1 extends ApiController implements Spa
             before(Routes.Pipeline.UNLOCK_PATH, mimeType, apiAuthenticationHelper::checkPipelineGroupOperateOfPipelineOrGroupInURLUserAnd403);
             before(Routes.Pipeline.TRIGGER_OPTIONS_PATH, mimeType, apiAuthenticationHelper::checkPipelineGroupOperateOfPipelineOrGroupInURLUserAnd403);
             before(Routes.Pipeline.SCHEDULE_PATH, mimeType, apiAuthenticationHelper::checkPipelineGroupOperateOfPipelineOrGroupInURLUserAnd403);
-            before(Routes.Pipeline.INSTANCE_PATH, mimeType, apiAuthenticationHelper::checkPipelineGroupOperateOfPipelineOrGroupInURLUserAnd403);
 
             post(Routes.Pipeline.PAUSE_PATH, mimeType, this::pause);
             post(Routes.Pipeline.UNPAUSE_PATH, mimeType, this::unpause);
             post(Routes.Pipeline.UNLOCK_PATH, mimeType, this::unlock);
             get(Routes.Pipeline.TRIGGER_OPTIONS_PATH, mimeType, this::triggerOptions);
             post(Routes.Pipeline.SCHEDULE_PATH, mimeType, this::schedule);
-            get(Routes.Pipeline.INSTANCE_PATH, mimeType, this::getInstanceInfo);
         });
     }
 
@@ -134,17 +131,6 @@ public class PipelineOperationsControllerV1 extends ApiController implements Spa
         return renderHTTPOperationResult(result, req, res);
     }
 
-    String getInstanceInfo(Request request, Response response) throws IOException {
-        String pipelineName = request.params("pipeline_name");
-        Integer pipelineCounter = getCounterValue(request);
-        HttpOperationResult result = new HttpOperationResult();
-        PipelineInstanceModel pipelineInstance = pipelineHistoryService.findPipelineInstance(pipelineName, pipelineCounter, currentUsername(), result);
-        if (result.canContinue()) {
-            return writerForTopLevelObject(request, response, (outputWriter) -> PipelineInstanceModelRepresenter.toJSON(outputWriter, pipelineInstance));
-        }
-        return renderHTTPOperationResult(result, request, response);
-    }
-
     private PipelineScheduleOptions getScheduleOptions(Request req) {
         if (StringUtils.isBlank(req.body())) {
             return new PipelineScheduleOptions();
@@ -154,15 +140,4 @@ public class PipelineOperationsControllerV1 extends ApiController implements Spa
         return PipelineScheduleOptionsRepresenter.fromJSON(jsonReader);
     }
 
-    private Integer getCounterValue(Request request) {
-        try {
-            int counter = Integer.parseInt(request.params("pipeline_counter"));
-            if (counter < 1) {
-                throw new UnprocessableEntityException("The pipeline counter cannot be less than 1.");
-            }
-            return counter;
-        } catch (NumberFormatException ex) {
-            throw new UnprocessableEntityException("The pipeline counter should be an integer.");
-        }
-    }
 }
