+++ api/api-pipeline-operations-v1/src/test/groovy/com/thoughtworks/go/apiv1/pipelineoperations/representers/MaterialRevisionRepresenterTest.groovy
@@ -16,56 +16,12 @@
 
 package com.thoughtworks.go.apiv1.pipelineoperations.representers
 
-
 import com.thoughtworks.go.api.util.GsonTransformer
-import com.thoughtworks.go.helper.ModificationsMother
 import org.junit.jupiter.api.Test
 
-import static com.thoughtworks.go.api.base.JsonOutputWriter.jsonDate
-import static com.thoughtworks.go.api.base.JsonUtils.toObjectString
-import static com.thoughtworks.go.apiv1.pipelineoperations.representers.MaterialRevisionRepresenter.fromJSON
-import static net.javacrumbs.jsonunit.fluent.JsonFluentAssert.assertThatJson
 import static org.assertj.core.api.Assertions.assertThat
 
 class MaterialRevisionRepresenterTest {
-  @Test
-  void 'should render material revision with modifications'() {
-    def materialRevision = ModificationsMother.createHgMaterialRevisions().getMaterialRevision(0)
-
-    def expectedJSON = [
-      "changed"      : false,
-      "material"     : [
-        "id"         : -1,
-        "name"       : "hg-url",
-        "fingerprint": "4290e91721d0a0be34955725cfd754113588d9c27c39f9bd1a97c15e55832515",
-        "type"       : "Mercurial",
-        "description": "URL: hg-url"
-      ],
-      "modifications": [
-        [
-          "id"           : -1,
-          "revision"     : "9fdcf27f16eadc362733328dd481d8a2c29915e1",
-          "modified_time": jsonDate(materialRevision.getModification(0).modifiedTime),
-          "user_name"    : "user2",
-          "comment"      : "comment2",
-          "email_address": "email2"
-        ],
-        [
-          "id"           : -1,
-          "revision"     : "eef77acd79809fc14ed82b79a312648d4a2801c6",
-          "modified_time": jsonDate(materialRevision.getModification(1).modifiedTime),
-          "user_name"    : "user1",
-          "comment"      : "comment1",
-          "email_address": "email1"
-        ]
-      ]
-    ]
-
-    def actualJson = toObjectString({ MaterialRevisionRepresenter.toJSON(it, materialRevision) })
-
-    assertThatJson(actualJson).isEqualTo(expectedJSON)
-  }
-
   @Test
   void 'should deserialize json'() {
     def inputJson = [
@@ -74,7 +30,7 @@ class MaterialRevisionRepresenterTest {
     ]
     def jsonReader = GsonTransformer.instance.jsonReaderFrom(inputJson)
 
-    def materialForScheduling = fromJSON(jsonReader)
+    def materialForScheduling = MaterialRevisionRepresenter.fromJSON(jsonReader)
 
     assertThat(materialForScheduling.fingerprint).isEqualTo("some-random-fingerprint")
     assertThat(materialForScheduling.revision).isEqualTo("some-revision")
